chore(scion-rcp): publish artifacts via the Central Portal
 * OSSRH is end of life as of June 30, 2025
 * Also updates the Node version used by the releaase version validation action. Besides, use rollup for bundling.
 * Also bumps maven plugin versions

resolves issue #87
