+++ pom.xml
@@ -60,12 +60,12 @@
 
 		<tycho.version>3.0.4</tycho.version>
 		<jacoco.version>0.8.10</jacoco.version>
-		<gpg-plugin.version>3.0.1</gpg-plugin.version>
-		<flatten-maven-plugin.version>1.4.1</flatten-maven-plugin.version>
-		<scm-plugin.version>2.0.0</scm-plugin.version>
-		<release-plugin.version>3.0.0</release-plugin.version>
-		<deploy-plugin.version>3.1.1</deploy-plugin.version>
-		<javadoc-plugin.version>2.9.1</javadoc-plugin.version>
+		<gpg-plugin.version>3.2.7</gpg-plugin.version>
+		<flatten-maven-plugin.version>1.7.0</flatten-maven-plugin.version>
+		<scm-plugin.version>2.1.0</scm-plugin.version>
+		<release-plugin.version>3.1.1</release-plugin.version>
+		<javadoc-plugin.version>3.11.2</javadoc-plugin.version>
+		<central-publishing-maven-plugin.version>0.8.0</central-publishing-maven-plugin.version>
 
 		<eclipse.release>2023-09</eclipse.release>
 
@@ -268,21 +268,11 @@
 
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-deploy-plugin</artifactId>
-				<version>${deploy-plugin.version}</version>
-				<configuration>
-					<skip>true</skip>
-				</configuration>
-			</plugin>
-
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-release-plugin</artifactId>
 				<version>${release-plugin.version}</version>
 				<configuration>
 					<autoVersionSubmodules>true</autoVersionSubmodules>
-					<useReleaseProfile>false</useReleaseProfile>
-					<releaseProfiles>release</releaseProfiles>
+					<releaseProfiles>javadoc-generate,gpg-sign</releaseProfiles>
 					<pushChanges>false</pushChanges>
 					<preparationGoals>tycho-versions:update-eclipse-metadata scm:add</preparationGoals>
 					<completionGoals>tycho-versions:update-eclipse-metadata scm:add</completionGoals>
@@ -298,7 +288,19 @@
 					<excludes>**/target/**</excludes>
 				</configuration>
 			</plugin>
-
+			
+			<plugin>
+          		<groupId>org.sonatype.central</groupId>
+          		<artifactId>central-publishing-maven-plugin</artifactId>
+          		<version>${central-publishing-maven-plugin.version}</version>
+          		<extensions>true</extensions>
+          		<configuration>
+            		<publishingServerId>sonatype-central</publishingServerId>
+            		<deploymentName>scion-rcp-${project.version}</deploymentName>
+            		<excludeArtifacts>
+						ch.sbb.scion.rcp.coveragereport,ch.sbb.scion.rcp.feature,ch.sbb.scion.rcp.feature.test,ch.sbb.scion.rcp.microfrontend.app.demo,ch.sbb.scion.rcp.microfrontend.app.demo.product,ch.sbb.scion.rcp.microfrontend.e3.app.demo,ch.sbb.scion.rcp.microfrontend.e3.app.demo.product,ch.sbb.scion.rcp.microfrontend.e3.app.demo.test,ch.sbb.scion.rcp.microfrontend.targetplatform,ch.sbb.scion.rcp.swtbot.feature,rcp-microfrontend-test,scion-parent</excludeArtifacts>
+          		</configuration>
+        	</plugin>
 		</plugins>
 
 		<pluginManagement>
@@ -307,6 +309,18 @@
 					<groupId>org.apache.maven.plugins</groupId>
 					<artifactId>maven-javadoc-plugin</artifactId>
 					<version>${javadoc-plugin.version}</version>
+					<executions>
+						<execution>
+							<id>attach-javadocs</id>
+							<goals>
+								<goal>jar</goal>
+							</goals>
+						</execution>
+					</executions>
+					<configuration>
+						<!-- We don't use the Java Platform Module System -->
+						<legacyMode>true</legacyMode>
+					</configuration>
 				</plugin>
 				<plugin>
 					<groupId>org.codehaus.mojo</groupId>
